scripts/video_gen_tests/requirements.txt
PyJWT>=2.8.0
requests>=2.31.0
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
